strands-agent-example/fund-advisor-agent-strands-stream/requirements.txt
strands-agents>=0.1.3
strands-agents-tools>=0.1.2
requests==2.32.3
python-dotenv==1.1.0
akshare==1.16.95
mem0ai==0.1.101
boto3==1.37.38
opensearch-py==2.8.0
duckduckgo_search==8.0.2
duckdb==1.3.0
fastapi==0.115.12
botocore==1.37.38
awscli==1.38.38
requests-aws4auth==1.3.1
PyYAML==6.0.2
retrying==1.3.4
pandas==2.2.1
mcp>=0.1.0